feat: wrap approve connection with auth component
