+++ apps/extension/src/Approvals/Approvals.tsx
@@ -14,7 +14,7 @@ import { ApproveUpdateDefaultAccount } from "./ApproveUpdateDefaultAccount";
 import { ConfirmSignature } from "./ConfirmSignArbitrary";
 import { ConfirmSignLedgerTx } from "./ConfirmSignLedgerTx";
 import { ConfirmSignTx } from "./ConfirmSignTx";
-import { UnlockKeychain } from "./UnlockKeychain";
+import { WithAuth } from "./WithAuth";
 
 export enum Status {
   Completed,
@@ -53,10 +53,6 @@ export const Approvals: React.FC = () => {
     >
       <Routes>
         <Route
-          path={TopLevelRoute.UnlockKeychain}
-          element={<UnlockKeychain />}
-        />
-        <Route
           path={`${TopLevelRoute.ApproveSignTx}/:msgId/:accountType/:signer`}
           element={<ApproveSignTx details={details} setDetails={setDetails} />}
         />
@@ -70,7 +66,11 @@ export const Approvals: React.FC = () => {
         />
         <Route
           path={TopLevelRoute.ApproveConnection}
-          element={<ApproveConnection />}
+          element={
+            <WithAuth>
+              <ApproveConnection />
+            </WithAuth>
+          }
         />
         <Route
           path={TopLevelRoute.ApproveDisconnection}
